POCs/AI_BETTER_ME/backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.11.7
python-multipart==0.0.6
websockets==12.0
aiohttp==3.9.1

# LangChain and LangGraph Ecosystem (Latest versions)
langchain==0.3.27
langchain-openai==0.3.33
langchain-community==0.3.29
langchain-core==0.3.76
langchain-ollama==0.3.8
langchain-text-splitters==0.3.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27

# Vector Store and ML
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional dependencies for specialized agents
requests==2.32.5
httpx==0.28.1
typing-extensions==4.14.1
ollama==0.5.3
openai==1.107.2

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Additional useful packages already in your environment
pandas==2.3.1
SQLAlchemy==2.0.41
PyYAML==6.0.2
tiktoken==0.11.0